+++ test/interfaces.d/templates/ppp_format.js
@@ -1,6 +1,0 @@
-module.exports = `
-auto  dsl-provider
-iface  dsl-provider  inet  ppp
-pre-up  /bin/ip  link  set  eth0  up
-provider  dsl-provider
-`