+++ test/interfaces.d/templates/static_ip_with_multi_dns_separated_by_space.js
@@ -1,9 +1,0 @@
-module.exports = `
-auto  eth0
-allow-hotplug  eth0
-iface  eth0  inet  static
-  address  10.0.0.1
-  netmask  255.255.240.0
-  dns-nameservers  10.0.0.1 8.8.4.4
-  gateway  10.0.0.1
-`